Relocate request_explicitly_requests_script_execution (Phase 8 batch 11, #688)
Issue #688 (parent #680, Phase 8) batch 11: migrate the
`request_explicitly_requests_script_execution` answer-only-mode helper
from `turn.rs` to `photon_feedback_derive.rs`. The 2-clause AND
predicate is the last pure photon-adjacent helper still living in
`turn.rs`'s top range (lines 540-1745) called out by Issue #688.

Migrated as a free fn in `photon_feedback_derive.rs` (`pub(super)`):

* `request_explicitly_requests_script_execution(request: &str) -> bool`

`turn.rs` adjustments:

* Removed the fn definition (originally at lines 517-542).
* Added `request_explicitly_requests_script_execution` to the existing
  `use super::photon_feedback_derive::{...}` block alongside
  `build_rerun_prompt_hint_if_eligible`.
* Test mod `tests` rewrote its `super::request_explicitly_requests_script_execution`
  re-export entry to `super::super::photon_feedback_derive::*` via a
  new import line.

LOC delta:

* `turn.rs`: 28,833 → 28,805 (-28 LOC)
* `photon_feedback_derive.rs`: 1,162 → 1,193 (+31 LOC)
* Cumulative `turn.rs`: 39,489 → 28,805 (-10,684 LOC, -27.1%)

Phase 8 (#688) is effectively complete — `turn.rs` no longer hosts
any photon-feedback-derive helper enumerated in the Issue #688 scope
(types, derive core, rerun-trigger family, dispatch helpers,
context-pack pipeline, evaluate dispatch, injection message,
answer-only script-execution helper). The remaining photon
references in `turn.rs` are test-mod call sites only (already routed
via `super::super::photon_feedback_derive::*`).

`pub(super)` matching the original module-private visibility.

Verification:

* `cargo fmt`
* `cargo clippy --lib --all-targets -- -D warnings` clean
* `cargo test --lib` 3,114 passed / 0 failed
